Moved ActiveSpan to common-api
