+++ instrumentation/activity/src/main/java/io/opentelemetry/android/instrumentation/activity/ActivityTracer.java
@@ -5,15 +5,15 @@
 
 package io.opentelemetry.android.instrumentation.activity;
 
-import static io.opentelemetry.android.common.RumConstants.APP_START_SPAN_NAME;
-import static io.opentelemetry.android.common.RumConstants.SCREEN_NAME_KEY;
-import static io.opentelemetry.android.common.RumConstants.START_TYPE_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.APP_START_SPAN_NAME;
+import static io.opentelemetry.android.internal.tools.RumConstants.SCREEN_NAME_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.START_TYPE_KEY;
 
 import android.app.Activity;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import io.opentelemetry.android.common.ActiveSpan;
 import io.opentelemetry.android.instrumentation.activity.startup.AppStartupTimer;
+import io.opentelemetry.android.instrumentation.common.ActiveSpan;
 import io.opentelemetry.android.internal.services.visiblescreen.VisibleScreenService;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.Span;
